Added testing code to a storage file
Tested some code for extracting and intepreting the binary file's insrtuctions.
Had issues with LUI receiving an immediate value which should've been negative, but was positive.
Problem seems to just be a lack of sign extension, but then again, the RISCV cheat sheet says to bit shift by 12 to the *left* ("<<"), instead of to the *right* (">>") which means we wouldn't be sign extending...?
Need to look into it later I guess