Fixed bug in LUI
LUI is a U-type operation, which uses UNSIGNED immediate values.
I hadn't noticed that detail, so extractImmediate() would always interpret the immediate values as signed, causing LUI to load in negative values instead of positive.
Fixed now, by adding an argument to extractImmediate() to tell it which operation type the immediate is from ("U", "I", "S", "J", or "B").