Fixed STORE operation bug and starting on memory
We were using '1' for the case for STORE operations, but that was incorrect, causing STORE operations to never be executed.
Instead we now use the correct case of '2', so STORE operations now get executed as they should.
Also started working on implementing some memory stuff maybe, but that kinda requires me to work out how to do STORE operations, which is what I'm currently working on.