Create .gitignore and added LOAD and STORE
Added a gitignore to finally get rid of those .pyx files we keep getting everytime we run the code.

Added the implementations for LOAD and STORE operations to the actual working code, instead of just having it in seperate files waiting to be added.

Finally, tested the stack pointer as well and it worked just fine, but we forgot to add the initial values for both the stack pointer ('sp' or 'x2') and the global pointer ('gp' or 'x3'), so I added those as well.