Code seems to work well enough now.
Tested the task 4 tests, and they all seem to pass, from what I can tell.

Couldn't run t14 in Venus, though, so couldn't test that one.

I know for a fact that there's an issue with the "signedIntToHex()" function I made, in that it accepts a value and a bit-count and returns a reseult, even if the given value couldn't actually be expressed with the number of bits specified.
After trying to fix this somehow, and having the code fuck up on a task when there was no incorrect value, I decided to scrap this check as it hindered us in a way I couldn't figure out how to even understand, let alone deal with.