Fixed code not getting the file path correctly
Code now gets the file path to find the tests in correctly. The folder the Main.py file is located in is fetched automatically, so we no longer have to have out source code in a folder with a specific name, as long as the tests are located in the path ".\tests\task[1-4]\[test_name].bin".

Also removed a bunch of old, commented-out code in Reader.py, from before we made the Main.py- and runCode.py-files.